requirements.txt
# Core 3270 dependencies
py3270>=0.3.5
pyte>=0.8.0

# AI/LLM dependencies
anthropic>=0.3.0
openai>=1.0.0

# CLI and utilities
click>=8.1.0
colorama>=0.4.6
rich>=13.0.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Async support
asyncio-throttle>=1.0.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0